% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.r
\name{corTest}
\alias{corTest}
\title{Robust correlation test}
\usage{
corTest(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Pearson's correlation coefficient.
}
\description{
Correlation test robust to \code{NA}s in the data.
}
