% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delTemp.r
\name{delTemp}
\alias{delTemp}
\title{Delete temporary directory}
\usage{
delTemp(tempDir, defaultTempDir = "C:/ecology/!Scratch")
}
\arguments{
\item{tempDir}{Character, directory path to delete.}

\item{defaultTemp}{Character, default temporary directory (temp directory will be redirected here).}
}
\value{
Nothing (deletes a directory).
}
\description{
Delete temporary directory. Useful for running Maxent 3.3.3k or earlier because it creates many temporary files.
}
