% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.r
\name{gaussian}
\alias{gaussian}
\title{Standardized bivariate Gaussian function}
\usage{
gaussian(x1, x2 = 0, mu1 = 0, mu2 = 0, sigma1 = 1, sigma2 = 0,
  rho = 0, ...)
}
\arguments{
\item{x1}{Numeric, values of the first variable.}

\item{x2, }{Numeric, values of the second variable. Default is 0, in which case only response to \code{x1} is returned.}

\item{mu1}{Numeric, mean of \code{x1} in a Gaussian function. The maximum value of the function with respect to \code{x1} occurs at this value of \code{x1}.}

\item{mu2}{Numeric, mean of \code{x2} in a Gaussian function. The maximum value of the function with respect to \code{x2} occurs at this value of \code{x2}.}

\item{sigma1}{Numeric, standard deviation of \code{x1} in a Gaussian function. Scales the width of the response to \code{x1}.}

\item{sigma2}{Numeric, standard deviation of \code{x2} in a Gaussian function. Scales the width of the response to \code{x2}.}

\item{rho}{Numeric, strength of covariance between \code{x1} and \code{x2}.}

\item{...}{Other arguments (unused).}
}
\description{
Bivariate Gaussian function useful for simulating the response of a species to two environmental variables. The function is standardized so that it yields a value of 1 when the covariates are at their "optimal" values. The default values produce a univariate Gaussian function centered on 0 with unit variance.
}
