% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaImport.r
\name{iaImport}
\alias{iaImport}
\title{Test importance of interaction terms in Maxent or omniscient model}
\usage{
iaImport(model, vars, sim, perms, testPres, testContrast,
  predPres = NULL, predContrast = NULL, ...)
}
\arguments{
\item{model}{Either a MaxEnt object or function object.}

\item{vars}{Character vector, names of variables to test interactions of.}

\item{perms}{Positive integer, number of times to permute data.}

\item{testPres}{Data frame with environment at test presence sites.}

\item{testContrast}{Data frame with environment at test background or absence sites.}

\item{predPres}{Predictions at test presences (if \code{NULL} or not supplied then will be computed).}

\item{predContrast}{Predictions at test background or absence sites (if \code{NULL} or not supplied then will be computed).}

\item{...}{Arguments to pass to \code{predictModel} function.}
}
\value{
Data frame with values of AUC (area under the receiver-operator curve), CBI (Continuous Boyce Index), and the Pearson correlation coefficient between predictions from the model with data as-observed and predictions with each variable permuted in turn. Permutations an be performed before the values are multiplied or after they are multiplied.
}
\description{
Test importance of interaction terms in Maxent or omniscient model
}
