% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEvals.r
\name{loadEvals}
\alias{loadEvals}
\title{Load evaluation files and compile into a single data frame}
\usage{
loadEvals(outDir, algos = c("omniscient", "brt", "gam", "maxent"))
}
\arguments{
\item{outDir}{Character, name of the \emph{scenario} directory in which a folder named "evaluations" is found containing evaluation files.}

\item{algos}{Character, name of algorithm(s) for which to collate evaluation files. Examples: \code{c('omniscient', 'brt', 'gam', 'maxent')}.}
}
\value{
Data frame.
}
\description{
This is a helper function that loads multiple evaluation files and compiles them into a single data frame.
}
