% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.r
\name{logistic}
\alias{logistic}
\title{Bivariate logistic function}
\usage{
logistic(x1, x2 = 0, b0 = 0, b1 = 1, b2 = 0, b12 = 0, ...)
}
\arguments{
\item{x1}{Numeric, first variable.}

\item{x2}{Numeric, second variable. Default is 0 (makes the function respond only to \code{x1} so it is univariate).}

\item{b0}{Numeric, intercept. Default is 0.}

\item{b1}{Numeric, slope with respect to \code{x1}. Default is 1.}

\item{b2}{Numeric, slope with respect to \code{x2}. Default is 0.}

\item{b12}{Numeric, slope with respect to \code{x1 * x2}. Default is 0 (no interaction).}

\item{...}{Other arguments (not used).}
}
\value{
A logistic function.
}
\description{
The bivariate logistic function.
}
\examples{
x1 <- seq(-1, 1, by=0.01)
pr1 <- logistic(x1=x1, b1=1)
pr2 <- logistic(x1=x1, b1=2)
pr3 <- logistic(x1=x1, b1=3)
pr4 <- logistic(x1=x1, b0=-1, b1=3)

plot(x1, pr1, ylab='Logistic output', type='l', ylim=c(0, 1))
lines(x1, pr2, col='red')
lines(x1, pr3, col='blue')
lines(x1, pr4, col='green')
}
