% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainEvalModels.r
\name{mainEvalModels}
\alias{mainEvalModels}
\title{Evaluate predictor importance of models trained on simulated data}
\usage{
mainEvalModels(scenarioDir, algos = c("omniscient", "brt", "gam",
  "maxent", "rf"), type = c("multivariate", "reduced", "univariate"),
  iters = 1:100, perms = 30L, ia = TRUE, strat = FALSE,
  overwrite = TRUE, fileAppend = NULL, verbose = 1, ...)
}
\arguments{
\item{scenarioDir}{Character, path name of scenario directory. Inside this must be a subfolder named "!scenario data" which is populated using the \code{mainMakeData} function. Also inside this folder will be at least one subfolder named "multivariate", "reduced", or "univariate".}

\item{algos}{Character list of model algorithms to evaluate. Options include \code{omniscient}, \code{brt} (boosted regression trees), \code{gam} (generalized additive models), \code{glm} (generalized linear models), \code{maxent} (Maxent, using version 3.3.3k or before), or \code{rf} (random forests).}

\item{type}{Character, type of models to train. Options include \code{multivariate} (use all variables in \code{vars}, \code{reduced} (a series of models, each using all but one variable in \code{vars}), and/or \code{univariate} (a series of models, one per variable in \code{vars}).}

\item{iters}{Vector of positive integers, data iterations to evaluate.}

\item{perms}{Positive integer, number of permutations for permutation tests of variable importance. Default is 30.}

\item{ia}{Logical, if \code{TRUE} (default) evaluate the importance of interaction terms in Maxent (permutation importance), BRTs (native importance only) and OMNISCIENT (permutation importance) models.}

\item{strat}{Logical, if \code{TRUE} then in addition to other kinds of variable importance evaluations, also try evaluating importance using just a stratified random sample of background sites sampled from across the predicted distribution of the species. The predicted distribution is generated by the SDM algorithm. Stratification is performed first by stretching the predicted values between [0, 1] then defining strata by binning the values into 10 groups ([0, 0.1], (0.1, 0.2], (0.2, 0.3], ..., (0.9, 1]). The number of samples taken per stratum is equal to a tenth of the number of background sites used to train the SDM.}

\item{overwrite}{Logical, if \code{TRUE} then write over existing evaluation files. Default is \code{FALSE}.}

\item{fileAppend}{Either \code{NULL} (append nothing) or a character string appended to the end of each model results file.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Arguments to pass to \code{\link[enmSdmPredImport]{iaImport}}.}
}
\value{
Nothing (saves a data frame to disc).
}
\description{
This function is used evaluate the importance of variables in models trained on simulated data. Typical implementation is to use \code{mainMakeData} to create simulated data sets, then \code{mainTrainModels} to train SDMs on those data sets, then \code{mainEvalModels} to evaluate the models.
}
\seealso{
\code{\link[enmSdmPredImport]{mainMakeData}}, \code{\link[enmSdmPredImport]{mainTrainModels}}
}
