% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMakeData.r
\name{mainMakeData}
\alias{mainMakeData}
\title{Master function to create common sets of data for simulations}
\usage{
mainMakeData(geography, response, scenarioDir, numTrainPres = 200,
  numTestPres = 200, numBg = 10000, iters = 1:100, circle = FALSE,
  size = 1001, fileAppend = NULL, b0 = NA, b1 = NA, b2 = NA,
  b11 = NA, b12 = NA, mu1 = NA, sigma1 = NA, sigma2 = NA,
  rho = NA, overwrite = FALSE, verbose = 1, ...)
}
\arguments{
\item{geography}{A list of lists describing the simulated environmental layers.  Each sublist pertains to one later. See \code{\link[enmSdmPredImport]{genesis}} for details.}

\item{response}{A function describing the response of the species to the environment. This must be one of: \code{\link[enmSdmPredImport]{logistic}}, \code{\link[enmSdmPredImport]{logisticShift}}, or \code{\link[enmSdmPredImport]{gaussian}}.}

\item{scenarioDir}{Path name of scenario directory. A folder named "!scenario data" will be created within this directory and one file per iteration stored within.}

\item{numTrainPres}{Positive integer, number of training presences to locate.}

\item{numTestPres}{Positive integer, number of test presences to locate. The number of test absences will also be equal to this value.}

\item{numBg}{Positive integer, number of training and number of test background sites to locate.}

\item{iters}{Vector of positive integers, data iterations to generate.}

\item{circle}{Logical, if \code{FALSE} (default), all landscapes are square. If \code{TRUE} then landscapes are circular.}

\item{size}{Positive integer, size of landscape in number of cells on a side.}

\item{fileAppend}{Either \code{NULL} or a character string. If a character string then this is appended to the file name. if \code{NULL} (default), nothing is appended. File names will be as "sim XXX.RData" or sim XXX APPEND.RData" where "XXX" is the iteration number and APPEND the string in \code{fileAppend}.}

\item{b0}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic intercept. Default is \code{NA}.}

\item{b1}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is \code{NA}.}

\item{b2}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is c.}

\item{b11}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Left-right shift along variable. Default is \code{NA}.}

\item{b12}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Interaction term. Default is \code{NA}.}

\item{mu1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}

\item{sigma1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{sigma2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{rho}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Covariance term. Default is \code{NA}.}

\item{overwrite}{Logical, if \code{TRUE} then save over pre-existing data files. Default is \code{FALSE}.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{mu2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}

\item{Other}{arguments (unused).}
}
\value{
Nothing (saves data files to disc in a subdirectory of the scenario directory named "!scenario data").
}
\description{
This function is typically used to create multiple simulated data sets for further analysis. The data sets represent data that one would typically possess in for applications of species distribution models (i.e., training and test presences, test absences and background sites). Typical implementation is to use \code{mainMakeData} to create simulated data sets, then \code{mainTrainModels} to train SDMs on those data sets, then \code{mainEvalModels} to evaluate the models.
}
\seealso{
\code{\link[enmSdmPredImport]{mainTrainModels}}, \code{\link[enmSdmPredImport]{mainEvalModels}}
}
