% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTemp.r
\name{makeTemp}
\alias{makeTemp}
\title{Make temporary directory}
\usage{
makeTemp(path = "C:/ecology/!Scratch/_scratchDir_dontDelete")
}
\arguments{
\item{path}{Character name of path.}
}
\value{
Path name of temporary directory plus creates the directory and assigns the raster temporary directory to this directory.
}
\description{
Make a temporary directory.
}
