% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.r
\name{predictModel}
\alias{predictModel}
\title{Predict OMNISCIENT, Maxent, BRT, RF, and GAM model objects}
\usage{
predictModel(model, data, ...)
}
\arguments{
\item{model}{Model object.}

\item{data}{Data frame.}

\item{...}{Arguments to pass to \code{predict} function.}
}
\value{
Numeric vector.
}
\description{
Predict OMNISCIENT, Maxent, BRT, RF, and GAM model objects.
}
