% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictOmniscientPermIa.r
\name{predictOmniscientPermIa}
\alias{predictOmniscientPermIa}
\title{Predictions for OMNISCIENT model using permuted interaction terms}
\usage{
predictOmniscientPermIa(model, data, sim, permProd = TRUE,
  permProdRule = NULL, ...)
}
\arguments{
\item{model}{Omniscient model object (of class "function") representing a bivariate \code{\link[enmSdmPredImport]{logistic}} or bivariate \code{\link[enmSdmPredImport]{gaussian}} function.}

\item{data}{Data frame.}

\item{sim}{Raster layer object used to standardize output.}

\item{permProd}{Logical, if \code{TRUE} (default) then permute product features using variables T1 and T2 according to \code{permProdRule}. If \code{FALSE} then predict as normal (no permutation).}

\item{permProdRule}{Character or logical. Rule for how permutation of product features is applied. \code{'before'} ==> Permute individual variable values then calculate product. \code{'after'} ==> Calculate product then permute across these values. Ignored if permProd is \code{NULL}.}

\item{...}{Arguments to pass to model function(s).}
}
\value{
Numeric vector.
export
}
\description{
Predictions for OMNISCIENT model using permuted interaction terms.
}
