% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predImportEval.r
\name{predImportEval}
\alias{predImportEval}
\title{Evaluate predictor importance of models trained on simulated data}
\usage{
predImportEval(simDir, modelDir, evalDir, algos = c("omniscient",
  "bioclim", "brt", "gam", "glm", "maxent", "maxnet", "rf"),
  type = c("multivariate", "reduced", "univariate"), iters = 1:100,
  perms = 30L, ia = TRUE, overwrite = FALSE, fileFlag = NULL,
  userdata = NULL, verbose = 1, ...)
}
\arguments{
\item{simDir}{Character, path name of directory in which scenario data files are saved.}

\item{modelDir}{Character, path name of directory in which model files are saved.}

\item{evalDir}{Character, path name of directory to which to save evaluations.}

\item{algos}{Character list of model algorithms to evaluate. Options include \code{omniscient}, \code{brt} (boosted regression trees), \code{gam} (generalized additive models), \code{glm} (generalized linear models), \code{maxent} (Maxent, using version 3.3.3k or before), \code{maxnet} (Maxent, version 3.4 or higher), or \code{rf} (random forests).}

\item{type}{Character, type of models to train. Options include \code{multivariate} (use all variables in \code{vars}, \code{reduced} (a series of models, each using all but one variable in \code{vars}), and/or \code{univariate} (a series of models, one per variable in \code{vars}).}

\item{iters}{Vector of positive integers, data iterations to evaluate.}

\item{perms}{Positive integer, number of permutations for permutation tests of variable importance. Default is 30.}

\item{ia}{Logical, if \code{TRUE} (default) evaluate the importance of interaction terms in Maxent (permutation importance), BRTs (native importance only) and OMNISCIENT (permutation importance) models.}

\item{overwrite}{Logical, if \code{TRUE} (default) then write over existing evaluation files..}

\item{fileFlag}{Either \code{NULL} or a character string. If a character string then this is included in the simulated data file name and each model file name. If \code{NULL} (default), nothing is added, so file names will be as "model XXX.RData". If a character string, then the file name will be as "ALGORITHM FLAG model XXX.RData" where "XXX" is the iteration number, "FLAG" the string in \code{fileFlag}, and "ALGORITHM" the model algorithm name.}

\item{userdata}{Either \code{NULL} or a data frame with extra information a user desires to include in the evaluation data frame output by this function. Note that an argument named \code{userdata} can also be specified in the \code{predImportMakeData} function, which will be stored as part of the \code{sim} object created by that function. This \code{sim}-associated metadata will \emph{also} be included in the evaluation data frame.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Arguments to pass to \code{\link[enmSdmPredImport]{iaImport}}.}
}
\value{
Nothing (saves a data frame to disc).
}
\description{
This function is used evaluate the importance of variables in models trained on simulated data. Typical implementation is to use \code{predImportMakeData} to create simulated data sets, then \code{predImportTrainModels} to train SDMs on those data sets, then \code{predImportEval} to evaluate the models.
}
\seealso{
\code{\link[enmSdmPredImport]{predImportMakeData}}, \code{\link[enmSdmPredImport]{predImportTrainModels}}
}
