% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predImportMakeData.r
\name{predImportMakeData}
\alias{predImportMakeData}
\title{Master function to create common sets of data for simulations}
\usage{
predImportMakeData(
  geography,
  response,
  simDir,
  numTrainPres = 200,
  numTrainAbs = 200,
  numTestPres = 200,
  numTestAbs = 200,
  numBg = 10000,
  iters = 1:100,
  circle = FALSE,
  sizeNative = 1024,
  sizeResampled = NULL,
  fileFlag = NULL,
  userdata = NULL,
  b0 = NA,
  b1 = NA,
  b2 = NA,
  b11 = NA,
  b12 = NA,
  mu1 = NA,
  mu2 = mu2,
  sigma1 = NA,
  sigma2 = NA,
  rho = NA,
  overwrite = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{geography}{A list of lists describing the simulated environmental layers.  Each sublist pertains to one later. See \code{\link[enmSdmPredImport]{genesis}} for details.}

\item{response}{A function describing the response of the species to the environment. This must be one of: \code{\link[enmSdmPredImport]{logistic}}, \code{\link[enmSdmPredImport]{logisticShift}}, or \code{\link[enmSdmPredImport]{gaussian}}.}

\item{simDir}{Character, path name of directory in which scenario data files are saved.}

\item{numTrainPres}{Positive integer, number of training presences to locate.}

\item{numTrainAbs}{Positive integer, number of training absences to locate.}

\item{numTestPres}{Positive integer, number of test presences to locate.}

\item{numTestAbs}{Positive integer, number of test absences to locate.}

\item{numBg}{Positive integer, number of training and number of test background sites to locate.}

\item{iters}{Vector of positive integers, data iterations to generate.}

\item{circle}{Logical, if \code{FALSE} (default), all landscapes are square. If \code{TRUE} then landscapes are circular.}

\item{sizeNative}{Positive integer, size of landscape in number of cells on a side. This specifies the spatial resolution at which the species perceives the landscape. See \emph{Details}.}

\item{sizeResampled}{Positive integer, size of landscape in number of cells on a side. This specifies the spatial resolution at which environmental data for model calibration and evaluation is available to the "modeler". Note that resampling will thus change the environmental values of the training and test data. See \emph{Details}.}

\item{fileFlag}{Either \code{NULL} or a character string. If a character string then this is included in the simulated data file name and each model file name. If \code{NULL} (default), nothing is added, so file names will be as "model XXX.RData". If a character string, then the file name will be as "ALGORITHM FLAG model XXX.RData" where "XXX" is the iteration number, "FLAG" the string in \code{fileFlag}, and "ALGORITHM" the model algorithm name.}

\item{userdata}{Either \code{NULL} (default) or a 1-line data frame to be included as part of the data in the \code{sim} object. This will be included in the evaluation data frame generated by the \code{predImportEval} function. It is useful for specifying aspects of the simulation that are not recorded by default. This metadata will be included in the evaluation data frame created by the \code{prediMportEval} function so will be available for analysis.}

\item{b0}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic intercept. Default is \code{NA}.}

\item{b1}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is \code{NA}.}

\item{b2}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is c.}

\item{b11}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Left-right shift along variable. Default is \code{NA}.}

\item{b12}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Interaction term. Default is \code{NA}.}

\item{mu1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}

\item{mu2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}

\item{sigma1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{sigma2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{rho}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Covariance term. Default is \code{NA}.}

\item{overwrite}{Logical, if \code{TRUE} then save over pre-existing data files. Default is \code{FALSE}.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Other arguments (unused).}
}
\value{
Nothing (saves data files to disc).
}
\description{
This function is used to create multiple simulated data sets for further analysis. The simulated data represents data that one would typically need for modeling species distributions (i.e., training and test presences, test absences and background sites). Typical implementation is to use \code{predImportMakeData} to create simulated data sets, then \code{predImportTrainModels} to train SDMs on those data sets, then \code{predImportEval} to evaluate the models.
}
\details{
In addition to its many capabilities, this function can be used to examine the effects of differences in spatial resolution of the scale at which a species responds to the environment and the scale at which environmental data is available to the modeler. When \code{sizeResampled} is \code{NULL} then the response scale is the same as the scale of environmental data. But when \code{sizeResampled} is not \code{NULL} and different from \code{sizeNative} then the landscape will be resampled to the stated resolution before environmental calibration and evaluation data is extracted. However, presences and absences and background sites will be drawn from the distribution of the true probability of presence generated using the native resolution landscape. Thus resampling to a different resolution maybe (intentionally) "confusing" to model because it wil be presented with data that is not necessarily indicative of the observed state (presence/background).
}
\seealso{
\code{\link[enmSdmPredImport]{predImportTrainModels}}, \code{\link[enmSdmPredImport]{predImportEval}}
}
