% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predImportMakeDataRes.r
\name{predImportMakeDataRes}
\alias{predImportMakeDataRes}
\title{Master function to create common sets of data for simulations}
\usage{
predImportMakeDataRes(geography, response, simDirNative, simDirSampled,
  numTrainPres = 200, numTestPres = 200, numBg = 10000,
  iters = 1:100, circle = FALSE, sizeNative = 1024,
  sizeSampled = 256, filePrepend = NULL, b0 = NA, b1 = NA,
  b2 = NA, b11 = NA, b12 = NA, mu1 = NA, sigma1 = NA,
  sigma2 = NA, rho = NA, overwrite = FALSE, verbose = 1, ...)
}
\arguments{
\item{geography}{A list of lists describing the simulated environmental layers.  Each sublist pertains to one later. See \code{\link[enmSdmPredImport]{genesis}} for details.}

\item{response}{A function describing the response of the species to the environment. This must be one of: \code{\link[enmSdmPredImport]{logistic}}, \code{\link[enmSdmPredImport]{logisticShift}}, or \code{\link[enmSdmPredImport]{gaussian}}.}

\item{simDirNative}{Character, path name of directory in which scenario data files are saved. This directory will contain data at the spatial resolution at which species respond to the environment.}

\item{numTrainPres}{Positive integer, number of training presences to locate.}

\item{numTestPres}{Positive integer, number of test presences to locate. The number of test absences will also be equal to this value.}

\item{numBg}{Positive integer, number of training and number of test background sites to locate.}

\item{iters}{Vector of positive integers, data iterations to generate.}

\item{circle}{Logical, if \code{FALSE} (default), all landscapes are square. If \code{TRUE} then landscapes are circular.}

\item{sizeNative}{Positive integer, size of landscape (number of cells on a side) at the resolution at which the species responds to the environment. This can be the same as \code{sizeSampled} or larger or smaller.}

\item{sizeSampled}{Positive integer, size of landscape (number of cells on a side) at the simulation assumes environmental data is available.}

\item{filePrepend}{Either \code{NULL} or a character string. If a character string then this is pre-pended to the simulated data file name. If \code{NULL} (default), nothing is appended. File names will be as "sim XXX.RData" (no append) or "PREPEND sim XXX.RData" where "XXX" is the iteration number and PREPEND the string in \code{filePrepend}.}

\item{b0}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic intercept. Default is \code{NA}.}

\item{b1}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is \code{NA}.}

\item{b2}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Logistic slope. Default is c.}

\item{b11}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Left-right shift along variable. Default is \code{NA}.}

\item{b12}{Numeric, parameters for \code{\link[enmSdmPredImport]{logistic}} or \code{\link[enmSdmPredImport]{logisticShift}} function specified in the \code{response} argument (above). Interaction term. Default is \code{NA}.}

\item{mu1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}

\item{sigma1}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{sigma2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Standard deviation of variable. Default is \code{NA}.}

\item{rho}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Covariance term. Default is \code{NA}.}

\item{overwrite}{Logical, if \code{TRUE} then save over pre-existing data files. Default is \code{FALSE}.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Other arguments (unused).}

\item{simDirNative}{Character, path name of directory in which scenario data files are saved. This directory will contain data at the spatial resolution at which predictors are available (which may or may not correspond to the scale at which species respond to the environment).}

\item{mu2}{Numeric, parameters for \code{\link[enmSdmPredImport]{gaussian}} function specified in the \code{response} argument (above). Mean of variable. Default is \code{NA}.}
}
\value{
Nothing (saves data files to disc).
}
\description{
This function is used to create multiple simulated data sets for further analysis. The simulated data represents data that one would typically need for modeling species distributions (i.e., training and test presences, test absences and background sites). Typical implementation is to use \code{predImportMakeData} to create simulated data sets, then \code{predImportTrainModels} to train SDMs on those data sets, then \code{predImportEval} to evaluate the models. This particular function is similar to \code{predImportMakeData} except that it allows simulating the effects of differences in spatial resolution between the scales at which the species responds to the environment and the scales at which predictor variables are available. See \emph{Details} for more information.
}
\details{
This function is useful for examining how well SDMs can measure the importance of variables if the spatial resolution (grain size) of the predictors is different from the spatial scale at which a species responds to the environment. Hereafter, the response scale is called the "native" scale and the scale at which predictors are available the "sampled" scale.  The sampled scale resolution can be larger or smaller than the native scale resolution.  
Unlike the \code{predImportMakeData} function, this function creates two \code{sim} objects and saves them in two directories. The first \code{sim} object, saved in the \code{simDirNative} will have training and test data for the native scale. The second \code{sim} object will have training and test data at the sampled scale from the same sites used to generate train and test data at the native scale. Generating two \code{sim} objects using the same underlying sites allows you to compare models trained/tested on the native versus sampled scales.  
Rasters in the native and sampled landscapes have the same extent. Thus, even though we're talking about resolution (size of a cell relative to the size of the raster), we specify resolution by the number of cells along the side of the landscape.  Since all landscapes are simulated to have the same extent, specifying a larger number of cells causes the resolution to become higher (grain size smaller).  Thus, the user must specify a size of the landscape (number of cells on a side) at the native and sampled resolutions.
}
\seealso{
\code{\link[enmSdmPredImport]{predImportMakeData}}, \code{\link[enmSdmPredImport]{predImportTrainModels}}, \code{\link[enmSdmPredImport]{predImportEval}}
}
