% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predOmniProd.r
\name{predOmniProd}
\alias{predOmniProd}
\title{Predict OMNISCIENT Gaussian model object using permutation of linear terms}
\usage{
predOmniProd(x1, x2, mu1, mu2, sigma1, sigma2, rho, product, extreme)
}
\arguments{
\item{x1}{Numeric, values of the first variable.}

\item{x2, }{Numeric, values of the second variable.}

\item{mu1}{Numeric, mean of \code{x1} in a Gaussian function. The maximum value of the function with respect to \code{x1} occurs at this value of \code{x1}.}

\item{mu2}{Numeric, mean of \code{x2} in a Gaussian function. The maximum value of the function with respect to \code{x2} occurs at this value of \code{x2}.}

\item{sigma1}{Numeric, standard deviation of \code{x1} in a Gaussian function. Scales the width of the response to \code{x1}.}

\item{sigma2}{Numeric, standard deviation of \code{x2} in a Gaussian function. Scales the width of the response to \code{x2}.}

\item{rho}{Numeric, strength of covariance between \code{x1} and \code{x2}.}

\item{product}{Numeric vector representing pre-multiplied (and possibly permuted) product: \code{(x1 - mu1) * (x2 - mu2)}.}

\item{extreme}{Numeric, most extreme value of product possible in this milieu. Used to standardize result to (0, 1].}
}
\value{
Numeric vector.
}
\description{
Predict OMNISCIENT Gaussian model object using permutation of linear terms (i.e., not interactions--see \code{\link[enmSdmPredImport]{predictOmniscientPermIa}}.
}
