% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichIsNaVec.r
\name{whichIsNaVec}
\alias{whichIsNaVec}
\title{Indices of NA values in one or more vectors.}
\usage{
whichIsNaVec(...)
}
\arguments{
\item{...}{One or more vectors.}
}
\value{
Integer vector.
}
\description{
Returns the indices of appearances of \code{NA} values in one or more vectors. The vectors don't have the be of same length, but the results may be less useful if they are not!
}
