% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTrainModels.r
\name{mainTrainModels}
\alias{mainTrainModels}
\title{Train SDMs across multiple iterations of a scenario}
\usage{
mainTrainModels(simDir, modelDir, vars, algos = c("omniscient", "brt",
  "gam", "glm", "maxent", "rf"), type = c("multivariate", "reduced",
  "univariate"), iters = 1:100, numBg = NULL, filePrepend = NULL,
  overwrite = TRUE, tempDir = raster::rasterOptions()$tmpdir,
  verbose = 1, ...)
}
\arguments{
\item{simDir}{Character, path name of directory in which scenario data files are saved.}

\item{modelDir}{Character, path name of directory in which model files are saved. Depending on whether multivariate, reduced, and/or univariate models are trained, inside this folder will be subfolders named "multivariate", "reduced", and/or "univariate" followed by the name of the algorithm (e.g., "multivariate brt").}

\item{vars}{Character vector, names of variables to use in model training. These should match the names in the \code{geography} argument supplied to the \code{mainMakeData} function. See \code{\link[enmSdmPredImport]{genesis}} function for more details on \code{geography}.}

\item{algos}{Character list of model algorithms. Options include \code{omniscient}, \code{brt} (boosted regression trees), \code{gam} (generalized additive models), \code{glm} (generalized linear models), \code{maxent} (Maxent, using version 3.3.3k or before), or \code{rf} (random forests).}

\item{type}{Character, type of models to train. Options include \code{multivariate} (use all variables in \code{vars}, \code{reduced} (a series of models, each using all but one variable in \code{vars}), and/or \code{univariate} (a series of models, one per variable in \code{vars}).}

\item{iters}{Vector of positive integers, data iterations to train models for.}

\item{numBg}{Positive integer or vector of positive integers or \code{NULL}. This is the number of background sites to use when training the model. If this is \code{NULL} (default), then the number of background sites will be equal to the number of sites available in the "sim" object created by \code{mainMakeData}. If this is a single integer, then the background sites will be taken from the first \code{numBg} sites available in the "sim" object. If this is a vector, then it must be the same length as \code{algos}. Each algorithm will then be presented with the respective number of background sites.}

\item{filePrepend}{Either \code{NULL} or a character string. If a character string then this is \emph{pre}-pended to the simulated data file name and each model file name. If \code{NULL} (default), nothing is pre-pended. File names will be as "model XXX.RData" (no append) or "ALGORITHM PREPEND model XXX.RData" where "XXX" is the iteration number, "PREPEND" the string in \code{filePrepend}, and "ALGORITHM" the model algorithm name.}

\item{overwrite}{Logical, if \code{TRUE} (default) then overwrite existing model results files..}

\item{tempDir}{Character, path of temporary directory. Used to store ancillary modeling files generated by Maxent 3.3.3k and earlier. Not used for any other modeling algorithm. Maxent can generate \emph{a lot} of these files which eventually fill up a disc. By specifying this folder you can have some control over where they are saved and thus if they fill up a hard drive (e.g., send temp files to a fast secondary drive with lots of space).}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Other arguments to pass to "train~~~" functions in the \pkg{enmSdm} package.}
}
\value{
Nothing (writes models to disc).
}
\description{
This function trains species distribution models on simulate data. Typical implementation is to use \code{mainMakeData} to create simulated data sets, then \code{mainTrainModels} to train SDMs on those data sets, then \code{mainEvalModels} to evaluate the models.
}
\seealso{
\code{\link[enmSdmPredImport]{mainMakeData}}, \code{\link[enmSdmPredImport]{mainEvalModels}}
}
