% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy.r
\name{noisy}
\alias{noisy}
\title{Add noise to a layer of layer raster stack by swapping values}
\usage{
noisy(landscape, geography)
}
\arguments{
\item{landscape}{Raster stack object.}

\item{geography}{List of lists, each sublist represents a layer in \code{landscape}. See \code{\link[enmSdmPredImport]{genesis}}.}
}
\value{
Raster stack.
}
\description{
Adds noise to a layer of a raster stack by swapping values across cells.
}
\examples{
set.seed(123)
geog <- list(
	control=list(type='linear', min=-1, max=1),
	noisy2=list(type='linear', min=-1, max=1, noisy=TRUE, noisyp=0.2),
	noisy4=list(type='linear', min=-1, max=1, noisy=TRUE, noisyp=0.4),
	noisy6=list(type='linear', min=-1, max=1, noisy=TRUE, noisyp=0.6),
	noisy8=list(type='linear', min=-1, max=1, noisy=TRUE, noisyp=0.8)
)
land <- genesis(geog, size=201)
plot(land)
}
