% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predImportEval.r
\name{predImportEval}
\alias{predImportEval}
\title{Evaluate predictor importance of models trained on simulated data}
\usage{
predImportEval(simDir, modelDir, evalDir, algos = c("omniscient",
  "bioclim", "brt", "gam", "glm", "maxent", "maxnet", "rf"),
  type = c("multivariate", "reduced", "univariate"), iters = 1:100,
  perms = 30L, ia = TRUE, strat = FALSE, overwrite = FALSE,
  filePrepend = NULL, verbose = 1, ...)
}
\arguments{
\item{simDir}{Character, path name of directory in which scenario data files are saved.}

\item{modelDir}{Character, path name of directory in which model files are saved.}

\item{evalDir}{Character, path name of directory to which to save evaluations.}

\item{algos}{Character list of model algorithms to evaluate. Options include \code{omniscient}, \code{brt} (boosted regression trees), \code{gam} (generalized additive models), \code{glm} (generalized linear models), \code{maxent} (Maxent, using version 3.3.3k or before), \code{maxnet} (Maxent, version 3.4 or higher), or \code{rf} (random forests).}

\item{type}{Character, type of models to train. Options include \code{multivariate} (use all variables in \code{vars}, \code{reduced} (a series of models, each using all but one variable in \code{vars}), and/or \code{univariate} (a series of models, one per variable in \code{vars}).}

\item{iters}{Vector of positive integers, data iterations to evaluate.}

\item{perms}{Positive integer, number of permutations for permutation tests of variable importance. Default is 30.}

\item{ia}{Logical, if \code{TRUE} (default) evaluate the importance of interaction terms in Maxent (permutation importance), BRTs (native importance only) and OMNISCIENT (permutation importance) models.}

\item{strat}{Logical, if \code{TRUE} then in addition to other kinds of variable importance evaluations, also try evaluating importance using just a stratified random sample of background sites sampled from across the predicted distribution of the species. The predicted distribution is generated by the SDM algorithm. Stratification is performed first by stretching the predicted values between [0, 1] then defining strata by binning the values into 10 groups ([0, 0.1], (0.1, 0.2], (0.2, 0.3], ..., (0.9, 1]). The number of samples taken per stratum is equal to a tenth of the number of background sites used to train the SDM.}

\item{overwrite}{Logical, if \code{TRUE} (default) then write over existing evaluation files..}

\item{filePrepend}{Either \code{NULL} or a character string. If a character string then this is \emph{pre}-pended to the simulated data file name and each model file name and \emph{ap}pended to the evaluation file name. If \code{NULL} (default), nothing is ap/pre-pended. File names will be as "sim XXX.RData" (no append) or "PREPEND sim XXX.RData" where "XXX" is the iteration number and PREPEND the string in \code{filePrepend}.}

\item{verbose}{Numeric, if 0 then show minimal output, 1 more output, 2 even more, >2 all of it.}

\item{...}{Arguments to pass to \code{\link[enmSdmPredImport]{iaImport}}.}
}
\value{
Nothing (saves a data frame to disc).
}
\description{
This function is used evaluate the importance of variables in models trained on simulated data. Typical implementation is to use \code{predImportMakeData} to create simulated data sets, then \code{predImportTrainModels} to train SDMs on those data sets, then \code{predImportEval} to evaluate the models.
}
\seealso{
\code{\link[enmSdmPredImport]{predImportMakeData}}, \code{\link[enmSdmPredImport]{predImportTrainModels}}
}
